gh-3 increase version, re-include ubbagent image again
Unfortunately the versions of the scripts and the underlying Weaviate
version diverge now. The feedback from the GCP marketplace team is that
this is intended. I've made sure to adjust the release notes with a link
to the actual Weaviate version. This is now version 0.22.8 with
underlying Weaviate version 0.22.7 which might get very confusing to
users. :(